# Compute can_holser once per save to speed up `d:` filter from ~20s to ~5s