@moxian 2025-01-05
> Honestly, I think this is an improvement. I expect this menu to be "What could this item fit into" rather than "What does this item fit into right now."

If you think this can be confusing, consider rewording `Can be stored in:` -> `Can be stored in (while empty):`. Perhaps only conditionally, if the item has non-rigid pockets.
(Personally I don't have an opinion here, and either wording is fine for my tastes)