@Brambor 2025-01-05
Could be. I don't like the "Can be stored in", it is probably broken. It doesn't list enough things, it probably only lists clothing.

Compare Can be stored in the description on a plank and then search crafting menu for L:122cm. Some of those containers should be shown in "can be stored in", in my opinion.

I am writing that because I don't know what the original intent was so I don't want to change the sentence without investigating what the list actually represents.

That said, this PR doesn't care what it lists, it just caches it.