@Brambor 2025-01-05
Oh, I know now. Let's use the cache as a first line of defence, giving candidates. Then check these candidates. It will still reduce the number of items to check from like 8000 to between 0 and 40. A bit slower. But there will be no change in functionality.

Gonna sleep now.