@Brambor 2025-01-05
You are commenting the existing code, not the last idea, right?

With the last idea https://github.com/CleverRaven/Cataclysm-DDA/pull/78950#issuecomment-2571499816, we would get candidates that need checking.

I asked on Discord and what breaks the last idea are items, that can get smaller, not bigger. They then could fit into more containers than precomputed.