@Brambor 2025-01-05
Or can't we just tell if item is modified? If it is, don't use this cache.