@CLIDragon 2025-01-07
I don't believe so, as it currently incorrectly handles items that can be modified to be smaller. The gun is a clear example, but the police baton or collapsing mop would also be a problem.