@Brambor 2025-01-07
I am converting this to a draft to clarify that it is not ready (in addition to the unresolved conversations).

I updated the OP on what needs to be done.