@Brambor 2025-01-16
I thought about this and want to try a different approach (but I am still lazy to test): Sort all itype containers by their pockets (probably volume, let's say unit U), and store a vector of that. When asked "can be contained in", consider only containers with at least U space. Then perform the usual `can_contain` check for each to disqualify by other parameters.

I want to experimentally verify that sorting by volume is the best. Sorting by more parameters would lead to an intersection operation of those results, which might be slow. We will see.