@Brambor 2025-03-17
- Please see #80203. It replaces this PR. As it does more. I made it a new PR to make it more clear.